Split of basetool, const, merge, templates and utils
