+++ srcjs/swiss-army-knife-card.js
@@ -42,6 +42,23 @@ import { selectUnit } from '@formatjs/intl-utils';
 import { fireEvent, stateIcon } from 'custom-card-helpers';
 import { version } from '../package.json';
 
+import {
+  SCALE_DIMENSIONS,
+  SVG_DEFAULT_DIMENSIONS,
+  SVG_DEFAULT_DIMENSIONS_HALF,
+  SVG_VIEW_BOX,
+  FONT_SIZE,
+  clamp,
+  round,
+  angle360,
+  range,
+} from './const';
+
+import Merge from './merge';
+import Utils from './utils';
+import Templates from './templates';
+import BaseTool from './basetool';
+
 // Original injector is buggy. Use a patched version, and store this local...
 import * as SvgInjector from '../dist/SVGInjector.min.js'; // lgtm[js/unused-local-variable]
 
@@ -51,29 +68,6 @@ console.info(
   'color: white; font-weight: bold; background: dimgray',
 );
 
-// Set sizes:
-// If svg size is changed, change the font size accordingly.
-// These two are related ;-) For font-size, 1em = 1%
-const SCALE_DIMENSIONS = 2;
-const SVG_DEFAULT_DIMENSIONS = 200 * SCALE_DIMENSIONS;
-const SVG_DEFAULT_DIMENSIONS_HALF = SVG_DEFAULT_DIMENSIONS / 2;
-const SVG_VIEW_BOX = SVG_DEFAULT_DIMENSIONS;
-const FONT_SIZE = SVG_DEFAULT_DIMENSIONS / 100;
-
-// Clamp number between two values
-const clamp = (min, num, max) => Math.min(Math.max(num, min), max);
-
-// Round to nearest value
-const round = (min, num, max) => ((Math.abs(num - min) > Math.abs(max - num)) ? max : min);
-
-// Force angle between 0 and 360, or even more for angle comparisons!
-const angle360 = (start, angle, end) => ((start < 0 || end < 0) ? angle + 360 : angle);
-
-// Size or range given by two values
-const range = (value1, value2) => Math.abs(value1 - value2);
-
-// const radianToDegrees = (radian) => (-radian / (Math.PI / 180));
-
 // ++ Class ++++++++++
 
 /**
@@ -82,300 +76,15 @@ const range = (value1, value2) => Math.abs(value1 - value2);
  */
 
 /**
- * Performs a deep merge of objects and returns new object. Does not modify
- * objects (immutable) and merges arrays via concatenation and filtering.
- *
- * @param {...object} objects - Objects to merge
- * @returns {object} New object with merged key/values
+  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
-class Merge {
-  static mergeDeep(...objects) {
-    const isObject = (obj) => obj && typeof obj === 'object';
-    return objects.reduce((prev, obj) => {
-      Object.keys(obj).forEach((key) => {
-        const pVal = prev[key];
-        const oVal = obj[key];
-        if (Array.isArray(pVal) && Array.isArray(oVal)) {
-          /* eslint no-param-reassign: 0 */
-          // Only if pVal is empty???
-
-          // #TODO:
-          // Should check for .id to match both arrays ?!?!?!?!
-          // Only concat if no ID or match found, otherwise mergeDeep ??
-          //
-          // concatenate and then reduce/merge the array based on id's if present??
-          //
-          prev[key] = pVal.concat(...oVal);
-        } else if (isObject(pVal) && isObject(oVal)) {
-          prev[key] = this.mergeDeep(pVal, oVal);
-        } else {
-          prev[key] = oVal;
-        }
-      });
-      return prev;
-    }, {});
-  }
-}
 
 /**
   +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
   +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
 
-/** ***************************************************************************
-  * Utils class
-  *
-  * Summary.
-  *
-  */
-
-class Utils {
-  /**
-  * Utils::calculateValueBetween()
-  *
-  * Summary.
-  * Clips the val value between start and end, and returns the between value ;-)
-  * Returned value is a fractional value between 0 and 1.
-  *
-  * Note 1:
-  * At start, state values are set to 'null' to make sure it has no value!
-  * If such a value is detected, return 0(%) as the relative value.
-  * In normal cases, this happens to be the _valuePrev, so 0% is ok!!!!
-  *
-  * Note 2:
-  * !xyz checks for "", null, undefined, false and number 0
-  * An extra check for NaN guards the result of this function ;-)
-  */
-
-  static calculateValueBetween(argStart, argEnd, argVal) {
-    // Check for valid argVal values and return 0 if invalid.
-    if (isNaN(argVal)) return 0;
-    if (!argVal) return 0;
-
-    // Valid argVal value: calculate fraction between 0 and 1
-    return (Math.min(Math.max(argVal, argStart), argEnd) - argStart) / (argEnd - argStart);
-  }
-
-  /**
-  * Utils::calculateSvgCoordinate()
-  *
-  * Summary.
-  * Calculate own (tool/tool) coordinates relative to centered toolset position.
-  * Tool coordinates are %
-  *
-  * Group is 50,40. Say SVG is 200x200. Group is 100,80 within 200x200.
-  * Tool is 10,50. 0.1 * 200 = 20 + (100 - 200/2) = 20 + 0.
-  */
-  static calculateSvgCoordinate(argOwn, argToolset) {
-    return (argOwn / 100) * (SVG_DEFAULT_DIMENSIONS)
-            + (argToolset - SVG_DEFAULT_DIMENSIONS_HALF);
-  }
-
-  /**
-  * Utils::calculateSvgDimension()
-  *
-  * Summary.
-  * Translate tool dimension like length or width to actual SVG dimension.
-  */
-
-  static calculateSvgDimension(argDimension) {
-    return (argDimension / 100) * (SVG_DEFAULT_DIMENSIONS);
-  }
-
-  static getLovelace() {
-    let root = document.querySelector('home-assistant');
-    root = root && root.shadowRoot;
-    root = root && root.querySelector('home-assistant-main');
-    root = root && root.shadowRoot;
-    root = root && root.querySelector('app-drawer-layout partial-panel-resolver, ha-drawer partial-panel-resolver');
-    root = (root && root.shadowRoot) || root;
-    root = root && root.querySelector('ha-panel-lovelace');
-    root = root && root.shadowRoot;
-    root = root && root.querySelector('hui-root');
-    if (root) {
-      const ll = root.lovelace;
-      ll.current_view = root.___curView;
-      return ll;
-    }
-    return null;
-  }
-}
-
-/**
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
- */
-
-/** ****************************************************************************
-  * Templates class
-  *
-  * Summary.
-  *
-  */
-
-class Templates {
-  /** ****************************************************************************
-  * Templates::replaceVariables()
-  *
-  * Summary.
-  * A toolset defines a template. This template is found and passed as argToolsetTemplate.
-  * This is actually a set of tools, nothing else...
-  * Also passed is the list of variables that should be replaced:
-  * - The list defined in the toolset
-  * - The defaults defined in the template itself, which are defined in the argToolsetTemplate
-  *
-  */
-
-  static replaceVariables3(argVariables, argTemplate) {
-    // If no variables specified, return template contents, not the first object, but the contents!
-    // ie template.toolset or template.colorstops. The toolset and colorstops objects are removed...
-    //
-    // If not, one gets toolsets[1].toolset.position iso toolsets[1].position.
-    //
-    if (!argVariables && !argTemplate.template.defaults) {
-      return argTemplate[argTemplate.template.type];
-    }
-    let variableArray = argVariables?.slice(0) ?? [];
-
-    // Merge given variables and defaults...
-    if (argTemplate.template.defaults) {
-      variableArray = variableArray.concat(argTemplate.template.defaults);
-    }
-
-    let jsonConfig = JSON.stringify(argTemplate[argTemplate.template.type]);
-    variableArray.forEach((variable) => {
-      const key = Object.keys(variable)[0];
-      const value = Object.values(variable)[0];
-      if (typeof value === 'number' || typeof value === 'boolean') {
-        const rxp2 = new RegExp(`"\\[\\[${key}\\]\\]"`, 'gm');
-        jsonConfig = jsonConfig.replace(rxp2, value);
-      }
-      if (typeof value === 'object') {
-        const rxp2 = new RegExp(`"\\[\\[${key}\\]\\]"`, 'gm');
-        const valueString = JSON.stringify(value);
-        jsonConfig = jsonConfig.replace(rxp2, valueString);
-      } else {
-        const rxp = new RegExp(`\\[\\[${key}\\]\\]`, 'gm');
-        jsonConfig = jsonConfig.replace(rxp, value);
-      }
-    });
-
-    return (JSON.parse(jsonConfig));
-  }
-
-  static getJsTemplateOrValueConfig(argTool, argValue) {
-    // Check for 'undefined' or 'null'
-    if (!argValue) return argValue;
-
-    // Check for primitive data types
-    if (['number', 'boolean', 'bigint', 'symbol'].includes(typeof argValue)) return argValue;
-
-    // We might have an object.
-    // Beware of the fact that this recursive function overwrites the argValue object,
-    // so clone argValue if this is the tool configuration...
-    if (typeof argValue === 'object') {
-      Object.keys(argValue).forEach((key) => {
-        argValue[key] = Templates.getJsTemplateOrValueConfig(argTool, argValue[key]);
-      });
-      return argValue;
-    }
-
-    // typeof should be a string now.
-    // The string might be a Javascript template surrounded by [[[<js>]]], or just a string.
-    const trimmedValue = argValue.trim();
-    if (trimmedValue.substring(0, 4) === '[[[[' && trimmedValue.slice(-4) === ']]]]') {
-      return Templates.evaluateJsTemplateConfig(argTool, trimmedValue.slice(4, -4));
-    } else {
-      // Just a plain string, return value.
-      return argValue;
-    }
-  }
-
-  static evaluateJsTemplateConfig(argTool, jsTemplate) {
-    try {
-      return new Function('tool_config', `'use strict'; ${jsTemplate}`).call(
-        this,
-        argTool,
-      );
-    } catch (e) {
-      e.name = 'Sak-evaluateJsTemplateConfig-Error';
-      throw e;
-    }
-  }
-  /** *****************************************************************************
-  * Templates::evaluateJsTemplate()
-  *
-  * Summary.
-  * Runs the JavaScript template.
-  *
-  * The arguments passed to the function are:
-  * - state, state of the current entity
-  * - states, the full array of states provided by hass
-  * - entity, the current entity and its configuration
-  * - user, the currently logged in user
-  * - hass, the hass object...
-  * - tool_config, the YAML configuration of the current tool
-  * - entity_config, the YAML configuration of configured entity in this tool
-  *
-  */
-
-  static evaluateJsTemplate(argTool, state, jsTemplate) {
-    try {
-      return new Function('state', 'states', 'entity', 'user', 'hass', 'tool_config', 'entity_config', `'use strict'; ${jsTemplate}`).call(
-        this,
-        state,
-        argTool._card._hass.states,
-        argTool.config.hasOwnProperty('entity_index') ? argTool._card.entities[argTool.config.entity_index] : undefined,
-        argTool._card._hass.user,
-        argTool._card._hass,
-        argTool.config,
-        argTool.config.hasOwnProperty('entity_index') ? argTool._card.config.entities[argTool.config.entity_index] : undefined,
-      );
-    } catch (e) {
-      e.name = 'Sak-evaluateJsTemplate-Error';
-      throw e;
-    }
-  }
-
-  /** *****************************************************************************
-  * Templates::getJsTemplateOrValue()
-  *
-  * Summary.
-  *
-  * References:
-  * - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Data_structures
-  * - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/typeof
-  *
-  */
-
-  static getJsTemplateOrValue(argTool, argState, argValue) {
-    // Check for 'undefined' or 'null'
-    if (!argValue) return argValue;
-
-    // Check for primitive data types
-    if (['number', 'boolean', 'bigint', 'symbol'].includes(typeof argValue)) return argValue;
-
-    // We might have an object.
-    // Beware of the fact that this recursive function overwrites the argValue object,
-    // so clone argValue if this is the tool configuration...
-    if (typeof argValue === 'object') {
-      Object.keys(argValue).forEach((key) => {
-        argValue[key] = Templates.getJsTemplateOrValue(argTool, argState, argValue[key]);
-      });
-      return argValue;
-    }
-
-    // typeof should be a string now.
-    // The string might be a Javascript template surrounded by [[[<js>]]], or just a string.
-    const trimmedValue = argValue.trim();
-    if (trimmedValue.substring(0, 3) === '[[[' && trimmedValue.slice(-3) === ']]]') {
-      return Templates.evaluateJsTemplate(argTool, argState, trimmedValue.slice(3, -3));
-    } else {
-      // Just a plain string, return value.
-      return argValue;
-    }
-  }
-}
 /**
   +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
   +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
@@ -693,611 +402,6 @@ class Toolset {
   +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
   +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
  */
-
-/** ***************************************************************************
-  * BaseTool class
-  *
-  * Summary.
-  *
-  */
-
-class BaseTool {
-  constructor(argToolset, argConfig, argPos) {
-    this.toolId = Math.random().toString(36).substr(2, 9);
-    this.toolset = argToolset;
-    this._card = this.toolset._card;
-    this.config = argConfig;
-
-    this.dev = { ...this._card.dev };
-
-    // The position is the absolute position of the GROUP within the svg viewport.
-    // The tool is positioned relative to this origin. A tool is always relative
-    // to a 200x200 default svg viewport. A (50,50) position of the tool
-    // centers the tool on the absolute position of the GROUP!
-    this.toolsetPos = argPos;
-
-    // Get SVG coordinates.
-    this.svg = {};
-
-    this.svg.cx = Utils.calculateSvgCoordinate(argConfig.position.cx, 0);
-    this.svg.cy = Utils.calculateSvgCoordinate(argConfig.position.cy, 0);
-
-    this.svg.height = argConfig.position.height ? Utils.calculateSvgDimension(argConfig.position.height) : 0;
-    this.svg.width = argConfig.position.width ? Utils.calculateSvgDimension(argConfig.position.width) : 0;
-
-    this.svg.x = (this.svg.cx) - (this.svg.width / 2);
-    this.svg.y = (this.svg.cy) - (this.svg.height / 2);
-
-    this.classes = {};
-    this.classes.card = {};
-    this.classes.toolset = {};
-    this.classes.tool = {};
-
-    this.styles = {};
-    this.styles.card = {};
-    this.styles.toolset = {};
-    this.styles.tool = {};
-
-    // Setup animation class and style and force initial processing by setting changed to true
-    this.animationClass = {};
-    this.animationClassHasChanged = true;
-
-    this.animationStyle = {};
-    this.animationStyleHasChanged = true;
-
-    // Process basic color stuff.
-    if (!this.config?.show?.style) {
-      if (!this.config.show) this.config.show = {};
-      this.config.show.style = 'default';
-    }
-    // Get colorstops and make a key/value store...
-    this.colorStops = {};
-    if ((this.config.colorstops) && (this.config.colorstops.colors)) {
-      Object.keys(this.config.colorstops.colors).forEach((key) => {
-        this.colorStops[key] = this.config.colorstops.colors[key];
-      });
-    }
-
-    if ((this.config.show.style === 'colorstop') && (this.config?.colorstops.colors)) {
-      this.sortedColorStops = Object.keys(this.config.colorstops.colors).map((n) => Number(n)).sort((a, b) => a - b);
-    }
-
-    this.csnew = {};
-    if ((this.config.csnew) && (this.config.csnew.colors)) {
-      this.config.csnew.colors.forEach((item, i) => {
-        this.csnew[item.stop] = this.config.csnew.colors[i];
-      });
-
-      this.sortedcsnew = Object.keys(this.csnew).map((n) => Number(n)).sort((a, b) => a - b);
-    }
-  }
-
-  /** *****************************************************************************
-  * BaseTool::textEllipsis()
-  *
-  * Summary.
-  * Very simple form of ellipsis, which is not supported by SVG.
-  * Cutoff text at number of characters and add '...'.
-  * This does NOT take into account the actual width of a character!
-  *
-  */
-  textEllipsis(argText, argEllipsis) {
-    if ((argEllipsis) && (argEllipsis < argText.length)) {
-      return argText.slice(0, argEllipsis - 1).concat('...');
-    } else {
-      return argText;
-    }
-  }
-
-  defaultEntityIndex() {
-    if (!this.default) {
-      this.default = {};
-      if (this.config.hasOwnProperty('entity_indexes')) {
-        this.default.entity_index = this.config.entity_indexes[0].entity_index;
-      } else {
-        // Must have entity_index! If not, just crash!
-        this.default.entity_index = this.config.entity_index;
-      }
-    }
-    return this.default.entity_index;
-  }
-
-  /** *****************************************************************************
-  * BaseTool::set value()
-  *
-  * Summary.
-  * Receive new state data for the entity this is linked to. Called from set hass;
-  *
-  */
-  set value(state) {
-    let localState = state;
-
-    if (this.dev.debug) console.log('BaseTool set value(state)', localState);
-    if (typeof (localState) !== 'undefined') if (this._stateValue?.toLowerCase() === localState.toLowerCase()) return false;
-
-    this.derivedEntity = null;
-
-    if (this.config.derived_entity) {
-      this.derivedEntity = Templates.getJsTemplateOrValue(this, state, Merge.mergeDeep(this.config.derived_entity));
-
-      localState = this.derivedEntity.state?.toString();
-    }
-
-    this._stateValuePrev = this._stateValue || localState;
-    this._stateValue = localState;
-    this._stateValueIsDirty = true;
-
-    // If animations defined, calculate style for current state.
-
-    // 2022.07.04 Temp disable this return, as animations should be able to process the 'undefined' state too!!!!
-    // if (this._stateValue == undefined) return;
-    // if (typeof(this._stateValue) === 'undefined') return;
-
-    let isMatch = false;
-    // #TODO:
-    // Modify this loop using .find() orso. It now keeps returning true for all items in animations list.
-    // It works, but can be more efficient ;-)
-
-    this.activeAnimation = null;
-
-    if (this.config.animations) Object.keys(this.config.animations).map((animation) => {
-      // NEW!!!
-      // Config more than 1 level deep is overwritten, so never changed after first evaluation. Stuff is overwritten???
-      const tempConfig = JSON.parse(JSON.stringify(this.config.animations[animation]));
-
-      const item = Templates.getJsTemplateOrValue(this, this._stateValue, Merge.mergeDeep(tempConfig));
-      // var item = Templates.getJsTemplateOrValue(this, this._stateValue, Merge.mergeDeep(this.config.animations[animation]));
-
-      if (isMatch) return true;
-
-      // #TODO:
-      // Default is item.state. But can also be item.custom_field[x], so you can compare with custom value
-      // Should index then not with item.state but item[custom_field[x]].toLowerCase() or similar...
-      // Or above, with the mapping of the item using the name?????
-
-      // Assume equals operator if not defined...
-      const operator = item.operator ? item.operator : '==';
-
-      switch (operator) {
-        case '==':
-          if (typeof (this._stateValue) === 'undefined') {
-            isMatch = (typeof item.state === 'undefined') || (item.state.toLowerCase() === 'undefined');
-          } else {
-            isMatch = this._stateValue.toLowerCase() === item.state.toLowerCase();
-          }
-          break;
-        case '!=':
-          if (typeof (this._stateValue) === 'undefined') {
-            isMatch = (item.state.toLowerCase() !== 'undefined');
-          } else {
-            isMatch = this._stateValue.toLowerCase() !== item.state.toLowerCase();
-          }
-          break;
-        case '>':
-          if (typeof (this._stateValue) !== 'undefined')
-            isMatch = Number(this._stateValue.toLowerCase()) > Number(item.state.toLowerCase());
-          break;
-        case '<':
-          if (typeof (this._stateValue) !== 'undefined')
-            isMatch = Number(this._stateValue.toLowerCase()) < Number(item.state.toLowerCase());
-          break;
-        case '>=':
-          if (typeof (this._stateValue) !== 'undefined')
-            isMatch = Number(this._stateValue.toLowerCase()) >= Number(item.state.toLowerCase());
-          break;
-        case '<=':
-          if (typeof (this._stateValue) !== 'undefined') {
-            isMatch = Number(this._stateValue.toLowerCase()) <= Number(item.state.toLowerCase());
-          }
-          break;
-        default:
-          // Unknown operator. Just do nothing and return;
-          isMatch = false;
-      }
-      if (this.dev.debug) console.log('BaseTool, animation, match, value, config, operator', isMatch, this._stateValue, item.state, item.operator);
-      if (!isMatch) return true;
-
-      if (!this.animationClass || !item.reuse) { this.animationClass = {}; }
-      if (item.classes) {
-        this.animationClass = Merge.mergeDeep(this.animationClass, item.classes);
-      }
-
-      if (!this.animationStyle || !item.reuse) this.animationStyle = {};
-      if (item.styles) {
-        this.animationStyle = Merge.mergeDeep(this.animationStyle, item.styles);
-      }
-
-      this.animationStyleHasChanged = true;
-
-      // #TODO:
-      // Store activeAnimation. Should be renamed, and used for more purposes, as via this method
-      // you can override any value from within an animation, not just the css style settings.
-      this.item = item;
-      this.activeAnimation = item;
-    });
-
-    return true;
-  }
-
-  /** *****************************************************************************
-  * BaseTool::set values()
-  *
-  * Summary.
-  * Receive new state data for the entity this is linked to. Called from set hass;
-  *
-  */
-
-  getEntityIndexFromAnimation(animation) {
-    // Check if animation has entity_index specified
-    if (animation.hasOwnProperty('entity_index')) return animation.entity_index;
-
-    // We need to get the default entity.
-    // If entity_index defined use that one...
-    if (this.config.hasOwnProperty('entity_index')) return this.config.entity_index;
-
-    // If entity_indexes is defined, take the
-    // first entity_index in the list as the default entity_index to use
-    if (this.config.entity_indexes) return (this.config.entity_indexes[0].entity_index);
-  }
-
-  getIndexInEntityIndexes(entityIdx) {
-    return this.config.entity_indexes.findIndex((element) => element.entity_index === entityIdx);
-  }
-
-  stateIsMatch(animation, state) {
-    let isMatch;
-    // NEW!!!
-    // Config more than 1 level deep is overwritten, so never changed after first evaluation. Stuff is overwritten???
-    const tempConfig = JSON.parse(JSON.stringify(animation));
-
-    const item = Templates.getJsTemplateOrValue(this, state, Merge.mergeDeep(tempConfig));
-
-    // Assume equals operator if not defined...
-    const operator = item.operator ? item.operator : '==';
-
-    switch (operator) {
-      case '==':
-        if (typeof (state) === 'undefined') {
-          isMatch = (typeof item.state === 'undefined') || (item.state.toLowerCase() === 'undefined');
-        } else {
-          isMatch = state.toLowerCase() === item.state.toLowerCase();
-        }
-        break;
-      case '!=':
-        if (typeof (state) === 'undefined') {
-          isMatch = (typeof item.state !== 'undefined') || (item.state.toLowerCase() !== 'undefined');
-        } else {
-          isMatch = state.toLowerCase() !== item.state.toLowerCase();
-        }
-        break;
-      case '>':
-        if (typeof (state) !== 'undefined')
-          isMatch = Number(state.toLowerCase()) > Number(item.state.toLowerCase());
-        break;
-      case '<':
-        if (typeof (state) !== 'undefined')
-          isMatch = Number(state.toLowerCase()) < Number(item.state.toLowerCase());
-        break;
-      case '>=':
-        if (typeof (state) !== 'undefined')
-          isMatch = Number(state.toLowerCase()) >= Number(item.state.toLowerCase());
-        break;
-      case '<=':
-        if (typeof (state) !== 'undefined')
-          isMatch = Number(state.toLowerCase()) <= Number(item.state.toLowerCase());
-        break;
-      default:
-        // Unknown operator. Just do nothing and return;
-        isMatch = false;
-    }
-    return isMatch;
-  }
-
-  mergeAnimationData(animation) {
-    if (!this.animationClass || !animation.reuse) this.animationClass = {};
-    if (animation.classes) {
-      this.animationClass = Merge.mergeDeep(this.animationClass, animation.classes);
-    }
-
-    if (!this.animationStyle || !animation.reuse) this.animationStyle = {};
-    if (animation.styles) {
-      this.animationStyle = Merge.mergeDeep(this.animationStyle, animation.styles);
-    }
-
-    this.animationStyleHasChanged = true;
-
-    // With more than 1 matching state (more entities), we have to preserve some
-    // extra data, such as setting the icon, name, area, etc. HOW?? Merge??
-
-    if (!this.item) this.item = {};
-    this.item = Merge.mergeDeep(this.item, animation);
-    this.activeAnimation = { ...animation }; // Merge.mergeDeep(this.activeAnimation, animation);
-  }
-
-  set values(states) {
-    if (!this._lastStateValues) this._lastStateValues = [];
-    if (!this._stateValues) this._stateValues = [];
-
-    const localStates = [...states];
-
-    if (this.dev.debug) console.log('BaseTool set values(state)', localStates);
-
-    // Loop through all values...
-    // var state;
-    for (let index = 0; index < states.length; ++index) {
-      // state = states[index];
-
-      if (typeof (localStates[index]) !== 'undefined') if (this._stateValues[index]?.toLowerCase() === localStates[index].toLowerCase()) {} else {
-        // State has changed, process...
-
-        if (this.config.derived_entities) {
-          this.derivedEntities[index] = Templates.getJsTemplateOrValue(this, states[index], Merge.mergeDeep(this.config.derived_entities[index]));
-
-          localStates[index] = this.derivedEntities[index].state?.toString();
-        }
-      }
-
-      this._lastStateValues[index] = this._stateValues[index] || localStates[index];
-      this._stateValues[index] = localStates[index];
-      this._stateValueIsDirty = true;
-
-      var isMatch = false;
-
-      this.activeAnimation = null;
-
-      if (this.config.animations) Object.keys(this.config.animations.map((aniKey, aniValue) => {
-        const statesIndex = this.getIndexInEntityIndexes(this.getEntityIndexFromAnimation(aniKey));
-        isMatch = this.stateIsMatch(aniKey, states[statesIndex]);
-
-        //        console.log("set values, animations", aniKey, aniValue, statesIndex, isMatch, states);
-
-        if (isMatch) this.mergeAnimationData(aniKey);
-      }));
-    }
-    this._stateValue = this._stateValues[this.getIndexInEntityIndexes(this.defaultEntityIndex())];
-    this._stateValuePrev = this._lastStateValues[this.getIndexInEntityIndexes(this.defaultEntityIndex())];
-
-    return true;
-  }
-
-  EnableHoverForInteraction() {
-    const hover = (this.config.hasOwnProperty('entity_index') || (this.config?.user_actions?.tap_action));
-    this.classes.tool.hover = !!hover;
-  }
-
-  /** *****************************************************************************
-  * BaseTool::MergeAnimationStyleIfChanged()
-  *
-  * Summary.
-  * Merge changed animationStyle with configured static styles.
-  *
-  */
-  MergeAnimationStyleIfChanged(argDefaultStyles) {
-    if (this.animationStyleHasChanged) {
-      this.animationStyleHasChanged = false;
-      if (argDefaultStyles) {
-        this.styles = Merge.mergeDeep(argDefaultStyles, this.config.styles, this.animationStyle);
-      } else {
-        this.styles = Merge.mergeDeep(this.config.styles, this.animationStyle);
-      }
-
-      if (this.styles.card) {
-        if (Object.keys(this.styles.card).length !== 0) {
-          this._card.styles.card = Merge.mergeDeep(this.styles.card);
-        }
-      }
-    }
-  }
-
-  /** *****************************************************************************
-  * BaseTool::MergeAnimationClassIfChanged()
-  *
-  * Summary.
-  * Merge changed animationclass with configured static styles.
-  *
-  */
-  MergeAnimationClassIfChanged(argDefaultClasses) {
-    // Hack
-    // @TODO This setting is still required for some reason. So this change is not detected...
-    this.animationClassHasChanged = true;
-
-    if (this.animationClassHasChanged) {
-      this.animationClassHasChanged = false;
-      if (argDefaultClasses) {
-        this.classes = Merge.mergeDeep(argDefaultClasses, this.config.classes, this.animationClass);
-      } else {
-        this.classes = Merge.mergeDeep(this.config.classes, this.animationClass);
-      }
-    }
-  }
-
-  /** *****************************************************************************
-  * BaseTool::MergeColorFromState()
-  *
-  * Summary.
-  * Merge color depending on state into colorStyle
-  *
-  */
-
-  MergeColorFromState(argStyleMap) {
-    if (this.config.hasOwnProperty('entity_index')) {
-      const color = this.getColorFromState(this._stateValue);
-      if (color !== '') {
-        argStyleMap.fill = this.config[this.config.show.style].fill ? color : '';
-        argStyleMap.stroke = this.config[this.config.show.style].stroke ? color : '';
-
-        // this.config[this.config.show.style].fill ? argStyleMap['fill'] = color : '';
-        // this.config[this.config.show.style].stroke ? argStyleMap['stroke'] = color : '';
-      }
-    }
-  }
-
-  /** *****************************************************************************
-  * BaseTool::MergeColorFromState2()
-  *
-  * Summary.
-  * Merge color depending on state into colorStyle
-  *
-  */
-
-  MergeColorFromState2(argStyleMap, argPart) {
-    if (this.config.hasOwnProperty('entity_index')) {
-      const fillColor = this.config[this.config.show.style].fill ? this.getColorFromState2(this._stateValue, argPart, 'fill') : '';
-      const strokeColor = this.config[this.config.show.style].stroke ? this.getColorFromState2(this._stateValue, argPart, 'stroke') : '';
-      if (fillColor !== '') {
-        argStyleMap.fill = fillColor;
-      }
-      if (strokeColor !== '') {
-        argStyleMap.stroke = strokeColor;
-      }
-    }
-  }
-
-  /** *****************************************************************************
-  * BaseTool::getColorFromState()
-  *
-  * Summary.
-  * Get color from colorstop or gradient depending on state.
-  *
-  */
-  getColorFromState(argValue) {
-    let color = '';
-    switch (this.config.show.style) {
-      case 'default':
-        break;
-      case 'fixedcolor':
-        color = this.config.color;
-        break;
-      case 'colorstop':
-      case 'colorstops':
-      case 'colorstopgradient':
-        color = this._card._calculateColor(argValue, this.colorStops, (this.config.show.style === 'colorstopgradient'));
-        break;
-      case 'minmaxgradient':
-        color = this._card._calculateColor(argValue, this.colorStopsMinMax, true);
-        break;
-    }
-    return color;
-  }
-
-  /** *****************************************************************************
-  * BaseTool::getColorFromState2()
-  *
-  * Summary.
-  * Get color from colorstop or gradient depending on state.
-  *
-  */
-  getColorFromState2(argValue, argPart, argProperty) {
-    let color = '';
-    switch (this.config.show.style) {
-      case 'colorstop':
-      case 'colorstops':
-      case 'colorstopgradient':
-        color = this._card._calculateColor2(argValue, this.csnew, argPart, argProperty, (this.config.show.style === 'colorstopgradient'));
-        break;
-      case 'minmaxgradient':
-        color = this._card._calculateColor2(argValue, this.colorStopsMinMax, argPart, argProperty, true);
-        break;
-    }
-    return color;
-  }
-
-  /** *****************************************************************************
-  * BaseTool::_processTapEvent()
-  *
-  * Summary.
-  * Processes the mouse click of the user and dispatches the event to the
-  * configure handler.
-  *
-  */
-
-  _processTapEvent(node, hass, config, actionConfig, entityId, parameterValue) {
-    let e;
-
-    if (!actionConfig) return;
-    fireEvent(node, 'haptic', actionConfig.haptic || 'medium');
-
-    if (this.dev.debug) console.log('_processTapEvent', config, actionConfig, entityId, parameterValue);
-    for (let i = 0; i < actionConfig.actions.length; i++) {
-      switch (actionConfig.actions[i].action) {
-        case 'more-info': {
-          if (typeof entityId !== 'undefined') {
-            e = new Event('hass-more-info', { composed: true });
-            e.detail = { entityId };
-            node.dispatchEvent(e);
-          }
-          break;
-        }
-        case 'navigate': {
-          if (!actionConfig.actions[i].navigation_path) return;
-          window.history.pushState(null, '', actionConfig.actions[i].navigation_path);
-          e = new Event('location-changed', { composed: true });
-          e.detail = { replace: false };
-          window.dispatchEvent(e);
-          break;
-        }
-        case 'call-service': {
-          if (!actionConfig.actions[i].service) return;
-          const [domain, service] = actionConfig.actions[i].service.split('.', 2);
-          const serviceData = { ...actionConfig.actions[i].service_data };
-
-          // Fill with current entity_id if none given
-          if (!serviceData.entity_id) {
-            serviceData.entity_id = entityId;
-          }
-          // If parameter defined, add this one with the parameterValue
-          if (actionConfig.actions[i].parameter) {
-            serviceData[actionConfig.actions[i].parameter] = parameterValue;
-          }
-          hass.callService(domain, service, serviceData);
-        }
-      }
-    }
-  }
-
-  /** *****************************************************************************
-  * BaseTool::handleTapEvent()
-  *
-  * Summary.
-  * Handles the first part of mouse click processing.
-  * It stops propagation to the parent and processes the event.
-  *
-  * The action can be configured per tool.
-  *
-  */
-
-  handleTapEvent(argEvent, argToolConfig) {
-    argEvent.stopPropagation();
-    argEvent.preventDefault();
-
-    let tapConfig;
-    // If no user_actions defined, AND there is an entity_index,
-    // define a default 'more-info' tap action
-    if (argToolConfig.hasOwnProperty('entity_index') && (!argToolConfig.user_actions)) {
-      tapConfig = {
-        haptic: 'light',
-        actions: [{
-          action: 'more-info',
-        }],
-      };
-    } else {
-      tapConfig = argToolConfig.user_actions?.tap_action;
-    }
-
-    if (!tapConfig) return;
-
-    this._processTapEvent(
-      this._card,
-      this._card._hass,
-      this.config,
-      tapConfig,
-      this._card.config.hasOwnProperty('entities')
-        ? this._card.config.entities[argToolConfig.entity_index]?.entity
-        : undefined,
-      undefined,
-    );
-  }
-} // -- CLASS
 
 /**
   +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
